IMC-728 update shipment doc
IMC-728 rubocop

IMC-728 update with active storage

IMC-728 update reviews

IMC-728 removed bp

IMC-728 admin shipment view aligned

IMC-728 client pipeline

IMC-728 rubocop


Former-commit-id: 9c51f13271f1a4c13b0a1a293f1c60daafbc3c45

+++ client/app/components/UserAccount/UserShipmentContent.jsx
@@ -114,31 +114,33 @@ class UserShipmentContent extends Component {
 
       Object.keys(uploadedDocs).forEach((key) => {
         docChecker[key] = true
-
-        docView.push(<div className={`flex-35 layout-row layout-align-start-start layout-padding ${adminStyles.uploaded_doc}`}>
-          <i className="fa fa-check flex-none" style={{ color: 'rgb(13, 177, 75)' }} />
-          <div className="layout-row flex layout-wrap" style={{ marginBottom: '12px' }}>
-            <h4 className="flex-100 layout-row">{documentTypes[key]}</h4>
-            {uploadedDocs[key].map(doc => (
-              <div className="flex-100 layout-row">
-                <a
-                  href={doc.signed_url}
-                  className={`${styles.eye_link} flex-none layout-row layout-align-center-center`}
-                  target="_blank"
-                >
-                  <i className="fa fa-eye pointy flex-none" />
-                </a>
-                <i
-                  className="fa fa-trash pointy flex-none"
-                  onClick={() => this.deleteDoc(doc)}
-                />
-                <p className="flex layout-row">
-                  {doc.text}
-                </p>
-              </div>
-            ))}
-          </div>
-        </div>)
+        
+        if (key !== 'shipment_recap') {
+          docView.push(<div className={`flex-35 layout-row layout-align-start-start layout-padding ${adminStyles.uploaded_doc}`}>
+            <i className="fa fa-check flex-none" style={{ color: 'rgb(13, 177, 75)' }} />
+            <div className="layout-row flex layout-wrap" style={{ marginBottom: '12px' }}>
+              <h4 className="flex-100 layout-row">{documentTypes[key]}</h4>
+              {uploadedDocs[key].map(doc => (
+                <div className="flex-100 layout-row">
+                  <a
+                    href={doc.signed_url}
+                    className={`${styles.eye_link} flex-none layout-row layout-align-center-center`}
+                    target="_blank"
+                  >
+                    <i className="fa fa-eye pointy flex-none" />
+                  </a>
+                  <i
+                    className="fa fa-trash pointy flex-none"
+                    onClick={() => this.deleteDoc(doc)}
+                  />
+                  <p className="flex layout-row">
+                    {doc.text}
+                  </p>
+                </div>
+              ))}
+            </div>
+          </div>)
+        }
       })
     }
